extract TimestampQueryBoxContainer; Use JoyUI Input instead of HTML input
