+++ src/components/MenuBar/TimestampQueryBox.tsx
@@ -1,12 +1,11 @@
 import {
     Box,
-    useColorScheme,
+    Input,
 } from "@mui/joy";
 
 import SearchIcon from "@mui/icons-material/Search";
 
 import useUiStore from "../../stores/uiStore";
-import {THEME_NAME} from "../../typings/config";
 import {UI_ELEMENT} from "../../typings/states";
 import {isDisabled} from "../../utils/states";
 import {updateWindowUrlHashParams} from "../../utils/url";
@@ -39,18 +38,13 @@ const handleTimestampQuery = (datetime: string) => {
  */
 const TimestampQueryBox = () => {
     const uiState = useUiStore((state) => state.uiState);
-    const {mode, systemMode} = useColorScheme();
     return (
         <Box className={"timestamp-query-box"}>
-            <input
+            <Input
                 disabled={isDisabled(uiState, UI_ELEMENT.NAVIGATION_BAR)}
                 id={"timestamp-query-box-input"}
-                step={"0.1"}
                 title={"Timestamp to seek to in UTC"}
                 type={"datetime-local"}
-                className={`timestamp-query-box-input ${(("system" === mode) ?
-                    systemMode :
-                    mode) ?? THEME_NAME.DARK}`}
                 defaultValue={new Date().toISOString()
                     .slice(0, -1)}
                 onKeyDown={(e) => {
