+++ src/components/MenuBar/TimestampQueryBoxContainer.css
@@ -1,0 +1,3 @@
+.timestamp-query-box-container {
+    display: flex;
+}
